pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "puppy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pyyaml", "jinja2", "markdown", "Pillow", "md2bbcode"]

[project.optional-dependencies]
test = ["pytest", "pyyaml"]

[project.scripts]
puppy = "puppy.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff.format]
quote-style = "single"
